requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    control run piptools.compile osscla
#
appdirs==1.4.2
authomatic==0.1.0.post1
boto3==1.4.4
botocore==1.5.6           # via boto3, pynamodb, s3transfer
certifi==2017.1.23
cffi==1.9.1
cryptography==1.7.1
docutils==0.13.1          # via botocore
enum34==1.1.6
flake8==2.4.0
flask-script==2.0.5
flask-sslify==0.1.5
flask==0.10.1             # via flask-script, flask-sslify
futures==3.1.1            # via s3transfer
gevent==1.2.1
greenlet==0.4.12
guard==1.0.1
gunicorn==19.7.1
idna==2.2
ipaddress==1.0.18
itsdangerous==0.24        # via flask
jinja2==2.9.4             # via flask
jmespath==0.9.1           # via boto3, botocore
markupsafe==0.23          # via jinja2
mccabe==0.3.1             # via flake8
ndg-httpsclient==0.4.2
pep8==1.5.7               # via flake8
py==1.4.32                # via pytest
pyasn1==0.1.9
pycparser==2.17
pyflakes==0.8.1           # via flake8
pygithub==1.35
pyjwt==1.5.2              # via pygithub
pynamodb==3.0.1
pyopenssl==16.2.0
pytest==2.8.5
python-dateutil==2.6.0
s3transfer==0.1.10        # via boto3
six==1.10.0
statsd==3.2.1
werkzeug==0.11.15         # via flask

# The following packages are considered to be unsafe in a requirements file:
# pip
